RenamingFiles: renaming some files.
